move local gen to fixtures
