+++ integrationtests/Cargo.toml
@@ -42,6 +42,7 @@ mint-client = { path = "../client/client-lib" }
 rand = "0.8"
 serde = { version = "1.0.149", features = [ "derive" ] }
 tokio = { version = "1.26.0", features = ["full"] }
+tokio-rustls = "0.23.4"
 tracing ="0.1.37"
 url = "2.3.1"
 hbbft = { git = "https://github.com/fedimint/hbbft" }
