+++ integrationtests/tests/fixtures/mod.rs
@@ -9,12 +9,14 @@ use std::sync::atomic::{AtomicI64, AtomicU16, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 
+use anyhow::Context;
 use bitcoin::hashes::{sha256, Hash};
 use bitcoin::{secp256k1, Address, KeyPair};
 use cln_rpc::ClnRpc;
 use fedimint_bitcoind::bitcoincore_rpc::{make_bitcoind_rpc, make_electrum_rpc, make_esplora_rpc};
 use fedimint_bitcoind::DynBitcoindRpc;
 use fedimint_client::module::gen::{ClientModuleGenRegistry, DynClientModuleGen};
+use fedimint_core::admin_client::PeerServerParams;
 use fedimint_core::api::WsFederationApi;
 use fedimint_core::bitcoin_rpc::read_bitcoin_backend_from_global_env;
 use fedimint_core::cancellable::Cancellable;
@@ -26,7 +28,7 @@ use fedimint_core::core::{
 use fedimint_core::db::mem_impl::MemDatabase;
 use fedimint_core::db::Database;
 use fedimint_core::module::registry::{ModuleDecoderRegistry, ModuleRegistry};
-use fedimint_core::module::DynServerModuleGen;
+use fedimint_core::module::{ApiAuth, DynServerModuleGen};
 use fedimint_core::outcome::TransactionStatus;
 use fedimint_core::server::DynServerModule;
 use fedimint_core::task::{timeout, RwLock, TaskGroup};
@@ -38,13 +40,13 @@ use fedimint_mint_client::MintClientGen;
 use fedimint_mint_server::common::db::NonceKeyPrefix;
 use fedimint_mint_server::common::MintOutput;
 use fedimint_mint_server::MintGen;
-use fedimint_server::config::{ConfigGenParams, ServerConfig};
+use fedimint_server::config::{gen_cert_and_key, ConfigGenParams, ServerConfig};
 use fedimint_server::consensus::{
     ConsensusProposal, FedimintConsensus, HbbftConsensusOutcome, TransactionSubmissionError,
 };
 use fedimint_server::db::GLOBAL_DATABASE_VERSION;
 use fedimint_server::net::connect::mock::{MockNetwork, StreamReliability};
-use fedimint_server::net::connect::{Connector, TlsTcpConnector};
+use fedimint_server::net::connect::{parse_host_port, Connector, TlsTcpConnector};
 use fedimint_server::net::peers::{DelayCalculator, PeerConnector};
 use fedimint_server::{consensus, EpochMessage, FedimintServer};
 use fedimint_testing::btc::fixtures::FakeBitcoinTest;
@@ -76,6 +78,7 @@ use rand::rngs::OsRng;
 use rand::RngCore;
 use real::{RealBitcoinTest, RealLightningTest};
 use tokio::sync::Mutex;
+use tokio_rustls::rustls;
 use tonic_lnd::connect;
 use tracing::{debug, info};
 use url::Url;
@@ -214,7 +217,7 @@ pub async fn fixtures(num_peers: u16, gateway_node: GatewayNode) -> anyhow::Resu
         bitcoin::network::constants::Network::Regtest,
         10,
     );
-    let params = ConfigGenParams::gen_local(&peers, base_port, "test", module_gens_params).unwrap();
+    let params = gen_local(&peers, base_port, "test", module_gens_params).unwrap();
 
     let server_module_inits = ServerModuleGenRegistry::from(vec![
         DynServerModuleGen::from(WalletGen),
@@ -458,6 +461,59 @@ pub async fn fixtures(num_peers: u16, gateway_node: GatewayNode) -> anyhow::Resu
     }
 
     Ok(fixtures)
+}
+
+/// config for servers running on different ports on our localhost
+pub fn gen_local(
+    peers: &[PeerId],
+    base_port: u16,
+    federation_name: &str,
+    modules: ServerModuleGenParamsRegistry,
+) -> anyhow::Result<HashMap<PeerId, ConfigGenParams>> {
+    let keys: HashMap<PeerId, (rustls::Certificate, rustls::PrivateKey)> = peers
+        .iter()
+        .map(|peer| {
+            let (cert, key) = gen_cert_and_key(&format!("peer-{}", peer.to_usize())).unwrap();
+            (*peer, (cert, key))
+        })
+        .collect::<HashMap<_, _>>();
+
+    let peer_params: BTreeMap<PeerId, PeerServerParams> = peers
+        .iter()
+        .map(|peer| {
+            let peer_port = base_port + u16::from(*peer) * 10;
+            let p2p_url = format!("ws://127.0.0.1:{peer_port}");
+            let api_url = format!("ws://127.0.0.1:{}", peer_port + 1);
+
+            let params: PeerServerParams = PeerServerParams {
+                cert: keys[peer].0.clone(),
+                p2p_url: p2p_url.parse().expect("Should parse"),
+                api_url: api_url.parse().expect("Should parse"),
+                name: format!("peer-{}", peer.to_usize()),
+            };
+            (*peer, params)
+        })
+        .collect();
+
+    peers
+        .iter()
+        .map(|peer| {
+            let bind_p2p = parse_host_port(peer_params[peer].clone().p2p_url)?;
+            let bind_api = parse_host_port(peer_params[peer].clone().api_url)?;
+
+            let params: ConfigGenParams = ConfigGenParams::new(
+                ApiAuth("dummy_password".to_string()),
+                bind_p2p.parse().context("when parsing bind_p2p")?,
+                bind_api.parse().context("when parsing bind_api")?,
+                keys[peer].1.clone(),
+                *peer,
+                peer_params.clone(),
+                federation_name.to_string(),
+                modules.clone(),
+            );
+            Ok((*peer, params))
+        })
+        .collect::<anyhow::Result<HashMap<_, _>>>()
 }
 
 pub async fn create_lightning_adapter(
